Création du dossier demineur/saves s'il n'existe pas
Bug que j'ai repéré durant certains tests, parce que je croyais que Java créait automatiquement un fichier et toute son arborescence s'il n'existait pas...

+++ app/src/main/java/fr/haronman/demineur/Sauvegarde.java
@@ -9,15 +9,20 @@ import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.io.Serializable;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.Optional;
 
 public class Sauvegarde implements Serializable{
-    public static final String CHEMIN_SAUVEGARDE = "C:\\Users\\hkoch\\Documents\\demineur_saves";
+    public static final String CHEMIN_SAUVEGARDE = "C:\\Users\\hkoch\\Documents\\demineur\\saves";
 
     public Sauvegarde(){}
 
     public void save(Partie partie) throws IOException, ClassNotFoundException{
         // Sauvegarde dans un fichier (resources/save)
+        File dossier = new File(CHEMIN_SAUVEGARDE);
+        if(!dossier.exists()){dossier.mkdirs();}
+
         String nomFichier = partie.getNomSave()+".save";
         File fichier = new File(Sauvegarde.CHEMIN_SAUVEGARDE+"\\"+nomFichier);
         if(!fichier.exists()){
@@ -31,41 +36,60 @@ public class Sauvegarde implements Serializable{
 
     public Optional<Partie> load(File fichier) throws IOException, ClassNotFoundException{
         // Chargement d'une partie dans un fichier
-        FileInputStream fis = new FileInputStream(fichier);
-        try(ObjectInputStream ois = new ObjectInputStream(fis)){
-            return Optional.ofNullable((Partie) ois.readObject());
+        File dossier = new File(CHEMIN_SAUVEGARDE);
+        if(!dossier.exists()){
+            dossier.mkdirs();
+            return Optional.empty();
+        }else{
+            FileInputStream fis = new FileInputStream(fichier);
+            try(ObjectInputStream ois = new ObjectInputStream(fis)){
+                return Optional.ofNullable((Partie) ois.readObject());
+            }
         }
     }
 
-    public static void delete(String nom){
-        File[] fichiers = new File(CHEMIN_SAUVEGARDE).listFiles();
-        if(fichiers != null || fichiers.length > 0){
-            for(File f : fichiers){
-                if(f.isFile() && f.getName().endsWith(".save")){
-                    String nomFichier = f.getName();
-                    nomFichier = nomFichier.substring(0, nomFichier.lastIndexOf("."));
-                    if(nomFichier.equals(nom)){
-                        f.delete();
-                        break;
+    public static void delete(String nom) throws IOException{
+        //Supression d'un fichier à partir de son nom
+        File dossier = new File(CHEMIN_SAUVEGARDE);
+        if(!dossier.exists()){
+            dossier.mkdirs();
+        }else{
+            File[] fichiers = new File(CHEMIN_SAUVEGARDE).listFiles();
+            if(fichiers != null || fichiers.length > 0){
+                for(File f : fichiers){
+                    if(f.isFile() && f.getName().endsWith(".save")){
+                        String nomFichier = f.getName();
+                        nomFichier = nomFichier.substring(0, nomFichier.lastIndexOf("."));
+                        if(nomFichier.equals(nom)){
+                            f.delete();
+                            break;
+                        }
                     }
                 }
             }
         }
     }
 
-    public static boolean nomDejaExistant(String nom){
-        File[] fichiers = new File(CHEMIN_SAUVEGARDE).listFiles();
-        if(fichiers != null || fichiers.length > 0){
-            for(File f : fichiers){
-                if(f.isFile() && f.getName().endsWith(".save")){
-                    String nomFichier = f.getName();
-                    nomFichier = nomFichier.substring(0, nomFichier.lastIndexOf("."));
-                    if(nomFichier.equals(nom)){
-                        return true;
+    public static boolean nomDejaExistant(String nom) throws IOException{
+        //Vérificationd e l'existence d'un fichier à partir d'un nom
+        File dossier = new File(CHEMIN_SAUVEGARDE);
+        if(!dossier.exists()){
+            dossier.mkdirs();
+            return false;
+        }else{
+            File[] fichiers = new File(CHEMIN_SAUVEGARDE).listFiles();
+            if(fichiers != null || fichiers.length > 0){
+                for(File f : fichiers){
+                    if(f.isFile() && f.getName().endsWith(".save")){
+                        String nomFichier = f.getName();
+                        nomFichier = nomFichier.substring(0, nomFichier.lastIndexOf("."));
+                        if(nomFichier.equals(nom)){
+                            return true;
+                        }
                     }
                 }
             }
+            return false;
         }
-        return false;
     }
 }
